# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PulseRecover API",
    "description": "AI Payment Reliability & Revenue Recovery Engine. Probabilistic AI proposes, deterministic policy decides, payment infrastructure executes, verification proves. All money amounts are integer paise (INR).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/merchant/sync": {
      "post": {
        "tags": [
          "merchant"
        ],
        "summary": "Post Sync",
        "operationId": "post_sync_api_v1_merchant_sync_post",
        "parameters": [
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 180,
              "minimum": 1,
              "default": 30,
              "title": "Window Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncRunView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "entity_counts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Entity Counts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "started_at",
          "status",
          "actor",
          "created_at"
        ],
        "title": "SyncRunView",
        "description": "POST /api/v1/merchant/sync \u2014 the sync_runs summary of the pass."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}